Template conversion (#2)
* Converted to using templates. There isn't a clean way to make a UE4 Plugin otherwise.  A typedef of WORLDSTATEPROPERTY_TYPE etc. is a bad idea because that makes the Plugin dependent on the main game code.

* make some BaseContext functions non-overrideable. These relate to the core state of a Context.

* - move headers into directories so that include paths must be specified from the root of the project directory. This keeps naming conflicts with other headers to a minimum by forcing a sort of "namespace"
- added some virtual destructors which UE4 complains about.

* - move pch back to root. It should not be included by consuming projects
- Moved STLTypes.h out of pch so that consumers are not forced to include pch.h

* remove all CPPs. IContext is a proper interface so there are no circular includes any more.

+++ Fluid-HTNCPP.UnitTests/BaseContextTests.cpp
@@ -61,7 +61,7 @@ namespace FluidHTNCPPUnitTests
             Assert::IsTrue(ctx.GetWorldStateChangeStack()[(int) DomainTestState::HasB].size() == 1);
             Assert::IsTrue(ctx.GetWorldStateChangeStack()[(int) DomainTestState::HasB].top().First() == EffectType::Permanent);
             Assert::IsTrue(ctx.GetWorldStateChangeStack()[(int) DomainTestState::HasB].top().Second() == 1);
-            Assert::IsTrue(ctx.GetWorldState().GetState((int) DomainTestState::HasB) == 0);
+            Assert::IsTrue(ctx.GetWorldState().GetState(DomainTestState::HasB) == 0);
         }
         TEST_METHOD(SetStateExecutingContext_ExpectedBehavior)
         {
@@ -73,7 +73,7 @@ namespace FluidHTNCPPUnitTests
 
             Assert::AreEqual(true, ctx.HasStateOneParam(DomainTestState::HasB));
             Assert::IsTrue(ctx.GetWorldStateChangeStack()[(int) DomainTestState::HasB].size() == 0);
-            Assert::IsTrue(ctx.GetWorldState().GetState((int) DomainTestState::HasB) == 1);
+            Assert::IsTrue(ctx.GetWorldState().GetState( DomainTestState::HasB) == 1);
         }
 
         TEST_METHOD(GetStatePlanningContext_ExpectedBehavior)
